requirements.txt
django>=5.2.1
djangogrpcframework>=0.2.1
djangorestframework>=3.16.0
djangorestframework_simplejwt>=5.5.0
grpcio>=1.71.0
grpcio-tools>= 1.71.0
psycopg2>=2.9.10
drf-spectacular>=0.28.0
celery>=5.5.2
redis>=6.1.0
requests>=2.32.3